apps/auth/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["ncgordeev <ncgordeev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.1.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = ">=5.5.0"
ipython = "==8.25"
pymongo = ">=4.11.2"
dnspython = ">=2.7.0"
python-dotenv = ">=1.0.1"
bcrypt = ">=4.3.0,<5.0.0"
pytest = ">=8.3.5,<9.0.0"
pytest-cov = ">=6.1.1,<7.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
flake8 = "^7.1.2"
testcontainers = "^4.10.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
addopts = "--reuse-db --cov=apps.users --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
markers = [
    "unit: тесты на единичные компоненты",
    "e2e: тесты, проверяющие поведение системы в целом"
]
python_files = ["test_*.py", "*_test.py", "*tests.py"]
testpaths = ["src"]